+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2026-01-03
+
+### Added
+
+- `dayOfYear`: Returns day of year (1-366).
+- `weekOfYear`: Returns ISO 8601 week number (1-53).
+
 ## [0.6.1] - 2025-12-30
 
 ### Documentation